+++ CHANGELOG.md
@@ -5,6 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- onFinish callback prop
+
+### Changed
+- unique package name to allow for npm publish
+- useTimer hook extracted from ProgressTimer
+- only publishing dist to npm, not src
+- components moved to components/
 
 ## [0.2.0] - 2022-06-02
 ### Added